+++ packages/react/src/components/Toast/styles.ts
@@ -1,0 +1,102 @@
+import * as Toast from '@radix-ui/react-toast'
+import { keyframes, styled } from '../../styles'
+
+export const ToastViewport = styled(Toast.Viewport, {
+  position: 'fixed',
+  bottom: 0,
+  right: 0,
+  display: 'flex',
+  flexDirection: 'column',
+  padding: 32,
+  gap: 10,
+  width: 360,
+  maxWidth: '100vw',
+  margin: 0,
+  listStyle: 'none',
+  zIndex: 2147483647,
+  outline: 'none',
+})
+
+export const hide = keyframes({
+  '0%': { opacity: 1 },
+  '100%': { opacity: 0 },
+})
+
+export const slideIn = keyframes({
+  from: { transform: `translateX(calc(100% + 32px))` },
+  to: { transform: 'translateX(0)' },
+})
+
+export const swipeOut = keyframes({
+  from: { transform: 'translateX(var(--radix-toast-swipe-end-x))' },
+  to: { transform: `translateX(calc(100% + 32px))` },
+})
+
+export const Header = styled('div', {
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  width: '100%',
+})
+
+export const ToastRoot = styled(Toast.Root, {
+  padding: '12px 20px',
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'flex-start',
+  gap: 4,
+
+  border: '1px solid $gray600',
+  borderRadius: '$sm',
+
+  background: '$gray800',
+  boxSizing: 'border-box',
+
+  '&[data-state="open"]': {
+    animation: `${slideIn} 150ms cubic-bezier(0.16, 1, 0.3, 1)`,
+  },
+  '&[data-state="closed"]': {
+    animation: `${hide} 100ms ease-in`,
+  },
+  '&[data-swipe="move"]': {
+    transform: 'translateX(var(--radix-toast-swipe-move-x))',
+  },
+  '&[data-swipe="cancel"]': {
+    transform: 'translateX(0)',
+    transition: 'transform 200ms ease-out',
+  },
+  '&[data-swipe="end"]': {
+    animation: `${swipeOut} 100ms ease-out`,
+  },
+})
+
+export const ToastTitle = styled(Toast.Title, {
+  fontFamily: '$default',
+  fontWeight: '$bold',
+  fontSize: '$xl',
+  lineHeight: '$base',
+  color: '$white',
+})
+
+export const ToastDescription = styled(Toast.Description, {
+  fontFamily: '$default',
+  fontWeight: '$regular',
+  fontSize: '$sm',
+  lineHeight: '$base',
+  color: '$gray200',
+})
+
+export const CloseButton = styled('button', {
+  all: 'unset',
+  boxSizing: 'border-box',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  cursor: 'pointer',
+  svg: {
+    color: '$gray200',
+  },
+  '&:hover svg': {
+    color: '$gray100',
+  },
+})