+++ packages/react/src/components/Tooltip/index.tsx
@@ -1,0 +1,27 @@
+import * as TooltipPrimitive from '@radix-ui/react-tooltip'
+import { ComponentProps, ReactNode } from 'react'
+import { TooltipArrow, TooltipContent } from './styles'
+
+interface TooltipPropType {
+  children: ReactNode
+  content: string
+}
+export function Tooltip({ children, content }: TooltipPropType) {
+  return (
+    <TooltipPrimitive.Provider delayDuration={200}>
+      <TooltipPrimitive.Root>
+        <TooltipPrimitive.Trigger asChild>{children}</TooltipPrimitive.Trigger>
+        <TooltipPrimitive.Portal>
+          <TooltipContent sideOffset={5}>
+            {content}
+            <TooltipArrow />
+          </TooltipContent>
+        </TooltipPrimitive.Portal>
+      </TooltipPrimitive.Root>
+    </TooltipPrimitive.Provider>
+  )
+}
+
+export interface TooltipProps extends ComponentProps<typeof Tooltip> {}
+
+Tooltip.displayName = 'Tooltip'