+++ packages/react/src/components/Tooltip/styles.ts
@@ -1,0 +1,52 @@
+import { keyframes, styled } from "../../styles";
+import * as Tooltip from "@radix-ui/react-tooltip";
+
+export const slideUpAndFade = keyframes({
+  "0%": { opacity: 0, transform: "translateY(2px)" },
+  "100%": { opacity: 1, transform: "translateY(0)" },
+});
+
+export const slideRightAndFade = keyframes({
+  "0%": { opacity: 0, transform: "translateX(-2px)" },
+  "100%": { opacity: 1, transform: "translateX(0)" },
+});
+
+export const slideDownAndFade = keyframes({
+  "0%": { opacity: 0, transform: "translateY(-2px)" },
+  "100%": { opacity: 1, transform: "translateY(0)" },
+});
+
+export const slideLeftAndFade = keyframes({
+  "0%": { opacity: 0, transform: "translateX(2px)" },
+  "100%": { opacity: 1, transform: "translateX(0)" },
+});
+
+export const TooltipContent = styled(Tooltip.Content, {
+  background: "$gray900",
+  display: "flex",
+  alignItems: "center",
+  padding: "12px 16px",
+  borderRadius: "$sm",
+  fontFamily: "$inter",
+  fontWeight: "$medium",
+  fontSize: "$sm",
+  lineHeight: "$short",
+  color: "$white",
+
+  userSelect: "none",
+  animationDuration: "400ms",
+  animationTimingFunction: "cubic-bezier(0.16, 1, 0.3, 1)",
+  willChange: "transform, opacity",
+  '&[data-state="delayed-open"]': {
+    '&[data-side="top"]': { animationName: slideDownAndFade },
+    '&[data-side="right"]': { animationName: slideLeftAndFade },
+    '&[data-side="bottom"]': { animationName: slideUpAndFade },
+    '&[data-side="left"]': { animationName: slideRightAndFade },
+  },
+});
+
+export const TooltipArrow = styled(Tooltip.Arrow, {
+  fill: "$gray900",
+  width: "16px",
+  height: "8px",
+});