+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @willry/tokens
 
+## 1.1.1
+
+### Patch Changes
+
+- Novos componentes para toast e tooltip
+
 ## 1.1.0
 
 ### Minor Changes